Group Name Changes

+++ resources/assets/js/integration/ws/chat/index.js
@@ -1,5 +1,5 @@
 import { store } from '../../../redux/Store';
-import { onNewChatAction, onNewMessageAction, onUpdateMessageAction, onDeleteMessagesAction, onDeleteChatAction, onMarkAsReadAction, onSelfDestructAction, onPublicKeyUpdatedAction, fetchChatsAction } from '../../../redux/actions/chatAction';
+import { onNewChatAction, onChatUpdatedAction, onNewMessageAction, onUpdateMessageAction, onDeleteMessagesAction, onDeleteChatAction, onMarkAsReadAction, onSelfDestructAction, onPublicKeyUpdatedAction, fetchChatsAction } from '../../../redux/actions/chatAction';
 import { onStatusChangedAction } from '../../../redux/actions/userAction';
 import { onConnectionStateChangedAction } from '../../../redux/actions/socketAction';
 import socket from '../../../common/socket';
@@ -45,6 +45,7 @@ export function monitorChats(userId) {
   subscription = socket.subscribe(`chat:${userId}`, event => {
     logger.log('CHAT', 'WS', `New "${event.type}" Event Received`, event.data);
     if (event.type === "chat:newChat") return store.dispatch(onNewChatAction(event.data, userId));
+    if (event.type === "chat:chatUpdated") return store.dispatch(onChatUpdatedAction(event.data, userId));
     if (event.type === "chat:newMessage") return store.dispatch(onNewMessageAction(event.data, userId));
     if (event.type === "chat:updateMessage") return store.dispatch(onUpdateMessageAction(event.data, userId));
     if (event.type === "chat:deleteMessages") return store.dispatch(onDeleteMessagesAction(event.data, userId));
